Rem Transition.ts's dep on area.ts
